requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.7
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
glfw==2.5.5
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyOpenGL @ file:///C:/Users/vnrju/Documents/TCC/venv/PyOpenGL-3.1.6-cp39-cp39-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Users/vnrju/Documents/TCC/venv/PyOpenGL_accelerate-3.1.6-cp39-cp39-win_amd64.whl
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.18.0
termcolor==2.1.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
